Refactor tests to check strong references count
In the daemon integration test, kaspad1 and kaspad2 now clone their arguments to avoid borrowing. Afterwards,  they check the number of strong references to the database, which should be one in each case, to ensure that no unexpected holders of references exist. Unecessary code was removed and refactored for clearer testing.