Add meta_db method to factory in consensusmanager
A meta_db method was added to the factory trait in consensusmanager. This method was added to provide access to the database through the factory, allowing direct checking of reference counts in testing integration for more precise memory management testing. It involved changes in daemon_integration_tests.rs and factory.rs. Also, the Kaspa-database was added as a dependency to consensusmanager's Cargo.toml.